Async cars

+++ app/Controllers/CarsController.js
@@ -12,38 +12,57 @@ function _drawCars() {
 export class CarsController {
   constructor() {
     ProxyState.on('cars', _drawCars)
+    carsService.getAllCars()
   }
   drawCars() {
     _drawCars()
     document.getElementById('modal-body-slot').innerHTML = getCarform()
   }
 
-  createCar() {
-    // prevents page reload
-    window.event.preventDefault()
-    console.log("submitted")
-    /** @type {HTMLFormElement} */
-    // @ts-ignore
-    const form = window.event.target
-    const carData = {
-      make: form.make.value,
-      model: form.model.value,
-      year: form.year.value,
-      price: form.price.value,
-      color: form.color.value,
-      description: form.description.value,
-      imgUrl: form.imgUrl.value
+  async createCar(id) {
+    try {
+      // prevents page reload
+      window.event.preventDefault()
+      console.log("submitted")
+      /** @type {HTMLFormElement} */
+      // @ts-ignore
+      const form = window.event.target
+      const carData = {
+        make: form.make.value,
+        model: form.model.value,
+        year: form.year.value,
+        price: form.price.value,
+        color: form.color.value,
+        description: form.description.value,
+        imgUrl: form.imgUrl.value
+      }
+      if (id == "undefined") {
+        await carsService.createCar(carData)
+      }
+      // else {
+      // NOTE NEED TO DO THIS FOR EDIT JUST LIKE LINE 38
+      // }
+      // clear form
+      form.reset()
+      // close modal
+      // @ts-ignore
+      bootstrap.Modal.getOrCreateInstance(document.getElementById('new-listing')).hide()
+    } catch (error) {
+      console.log(error.message)
     }
-    carsService.createCar(carData)
-    // clear form
-    form.reset()
-    // close modal
-    // @ts-ignore
-    bootstrap.Modal.getOrCreateInstance(document.getElementById('new-listing')).hide()
   }
 
-  removeCar(id) {
-    console.log('deleting', id)
-    carsService.removeCar(id)
+  async removeCar(id) {
+    try {
+      const foundCar = ProxyState.cars.find(c => c.id == id)
+      console.log('deleting', foundCar)
+      if (await confirmation(`Are you sure you want to delete ${foundCar.make}?`)) {
+        await carsService.removeCar(id)
+        // toast(`${foundcar.make} was removerd!`)
+      }
+    } catch (error) {
+      toast(error.message)
+      console.log(error.message)
+    }
   }
 }
